Add support for generating HOTP and MOTP tokens.

+++ src/otp_vault/main.py
@@ -9,19 +9,26 @@ from __future__ import annotations
 # Built-in Modules:
 import sys
 from collections.abc import Callable
-from typing import Optional
+from typing import Any, Literal, Optional
 
 # Third-party Modules:
 from tap import Tap
 
 # Local Modules:
-from . import __version__
+from . import __version__, otp
 from .clipboard import set_clipboard
-from .database import Database
-from .otp import totp
+from .database import Database, Secret
 
 
 DESCRIPTION: str = "An authenticator app."
+TOKEN_FUNCTIONS: tuple[str, ...] = ("hotp", "motp", "totp")
+
+
+def get_token(token_type: str, *args: Any, **kwargs: Any) -> str:
+	if token_type not in TOKEN_FUNCTIONS:
+		raise ValueError(f"{token_type} not in {TOKEN_FUNCTIONS}.")
+	otp_function: Callable[..., str] = getattr(otp, token_type)
+	return otp_function(*args, **kwargs)
 
 
 class ArgumentParser(Tap):  # pragma: no cover
@@ -31,6 +38,12 @@ class ArgumentParser(Tap):  # pragma: no cover
 	"""Change an existing password."""
 	add: Optional[str] = None
 	"""Add a secret key."""
+	type: Literal["hotp", "motp", "totp"] = "totp"
+	"""The type of OTP used (used with --add)."""
+	length: int = 6
+	"""The length of the output token (used with --add)."""
+	initial_input: str = "0"
+	"""The pin / counter / start time used as the moving factor (used with --add)."""
 	search: Optional[str] = None
 	"""Search for a secret key."""
 	copy: Optional[int] = None
@@ -81,6 +94,9 @@ class ArgumentParser(Tap):  # pragma: no cover
 		self.add_argument("password", metavar="password")
 		self.add_argument("--change-password", metavar="new_password")
 		self.add_argument("-a", "--add", nargs=2, metavar=("label", "key"))
+		self.add_argument("-t", "--type", metavar="type")
+		self.add_argument("-l", "--length", metavar="length")
+		self.add_argument("-i", "--initial-input", metavar="value")
 		self.add_argument("-s", "--search", metavar="text")
 		self.add_argument("-c", "--copy", metavar="item")
 		self.add_argument("-d", "--delete", metavar="item")
@@ -105,7 +121,14 @@ def change_password(database: Database, error_handler: Callable[[str], None], pa
 
 
 def add_secret(
-	database: Database, error_handler: Callable[[str], None], password: str, label: str, key: str
+	database: Database,
+	error_handler: Callable[[str], None],
+	password: str,
+	label: str,
+	key: str,
+	token_type: str,
+	length: int,
+	initial_input: str,
 ) -> None:
 	"""
 	Adds a secret to a database.
@@ -116,9 +139,12 @@ def add_secret(
 		password: The password for the database.
 		label: A label for the secret.
 		key: The key for the secret.
+		token_type: A valid OTP token type.
+		length: The desired length of the generated code.
+		initial_input: A moving factor value, such as MOTP pin, HOTP counter, or TOTP start time.
 	"""
 	try:
-		database.add_secret(password, label, key)
+		database.add_secret(password, label, key, token_type, length, initial_input)
 	except ValueError as e:
 		error_handler(str(e))
 	else:
@@ -147,48 +173,63 @@ def search_secrets(
 		delete: Delete the corresponding search result from the database.
 		update: Update the corresponding search result with a new label.
 	"""
-	results: tuple[tuple[str, str], ...] = database.search_secrets(text)
+	results: tuple[Secret, ...] = database.search_secrets(text)
 	if not results:
 		sys.exit("No results found.")  # Prints to STDERR and exits with status code 1.
 	for i, result in enumerate(results):
-		label, key = result
-		if copy is not None:
-			if not 1 <= copy <= len(results):
-				error_handler(f"Item {copy} to copy not in range 1-{len(results)}")
-				break
-			elif i + 1 == copy:
-				set_clipboard(totp(key))
+		label, key, token_type, length, initial_input = result
+		if copy is not None and not 1 <= copy <= len(results):
+			error_handler(f"Item {copy} to copy not in range 1-{len(results)}")
+			break
+		elif copy is not None and i + 1 == copy:
+			token: str = get_token(token_type, key, initial_input, length=length)
+			if token_type == "hotp":
+				# Increment the counter.
+				database.increment_initial_input(password, result, amount=1)
+			status: bool = set_clipboard(token)
+			if status:
 				print(f"Item {copy} ({label}) copied to clipboard.")
-				break
-		elif delete is not None:
-			if not 1 <= delete <= len(results):
-				error_handler(f"Item {delete} to delete not in range 1-{len(results)}")
-				break
-			elif i + 1 == delete:
-				database.delete_secret(password, label)
-				print(f"Item {delete} ({label}) deleted.")
-				break
-		elif update is not None:
-			# update is a list containing the item number and the new label.
-			if not 1 <= update[0] <= len(results):
-				error_handler(f"Item {update[0]} to update not in range 1-{len(results)}")
-				break
-			elif i + 1 == update[0]:
-				database.update_secret(password, label, update[1])
-				print(f"Item {update[0]} ({label}) updated to {update[1]}.")
-				break
+			else:
+				print(f"{label}: {token}")
+			break
+		elif delete is not None and not 1 <= delete <= len(results):
+			error_handler(f"Item {delete} to delete not in range 1-{len(results)}")
+			break
+		elif delete is not None and i + 1 == delete:
+			database.delete_secret(password, label)
+			print(f"Item {delete} ({label}) deleted.")
+			break
+		elif update is not None and not 1 <= update[0] <= len(results):
+			error_handler(f"Item {update[0]} to update not in range 1-{len(results)}")
+			break
+		elif update is not None and i + 1 == update[0]:
+			selected_item, new_label = update
+			database.update_secret(password, label, new_label)
+			print(f"Item {selected_item} ({label}) updated to {new_label}.")
+			break
 		else:
 			# Just searching.
-			print(f"{i + 1}: {label}, {totp(key)}")
+			print(f"{i + 1}: {label}")
 
 
 def main(parsed_args: ArgumentParser) -> None:  # pragma: no cover
 	database = Database(parsed_args.password)
 	if parsed_args.change_password is not None:
 		change_password(database, parsed_args.error, parsed_args.change_password)
+	elif parsed_args.type not in TOKEN_FUNCTIONS:
+		parsed_args.error(f"{parsed_args.type} not in {list(TOKEN_FUNCTIONS)}")
 	elif parsed_args.add is not None:
 		label, key = (i.strip() for i in parsed_args.add)
-		add_secret(database, parsed_args.error, parsed_args.password, label, key)
+		add_secret(
+			database,
+			parsed_args.error,
+			parsed_args.password,
+			label,
+			key,
+			parsed_args.type,
+			parsed_args.length,
+			parsed_args.initial_input,
+		)
 	elif parsed_args.search is not None:
 		search_secrets(
 			database,
